# lib/api-zod/src/generated/types/affiliateUrlEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface AffiliateUrlEntry {
  url: string;
  retailer: string;
}