# lib/api-zod/src/generated/types/shopGenerateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ShopGenerateRequest {
  /** warm, cool, or neutral */
  undertone: string;
  /** Spring, Summer, Autumn, or Winter */
  color_season?: string;
  style_archetype?: string;
  aesthetic_archetypes?: string[];
  skin_type?: string;
  shopping_keywords?: string[];
  /** Hex color codes from the analysis */
  color_palette?: string[];
  fashion_direction?: string;
  makeup_direction?: string;
}