# lib/api-zod/src/generated/types/shopProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ShopProductAffiliateUrls } from "./shopProductAffiliateUrls";
import type { ShopProductCategory } from "./shopProductCategory";
import type { ShopProductPriceTier } from "./shopProductPriceTier";

export interface ShopProduct {
  id: number;
  name: string;
  brand: string;
  retailer: string;
  category: ShopProductCategory;
  /** AI-written reason this product suits the profile (~40 words) */
  description: string;
  price_tier: ShopProductPriceTier;
  /** Dominant product colors as hex codes */
  color_hex: string[];
  /** Undertones this product suits (warm/cool/neutral) */
  undertones: string[];
  /** Color seasons this product suits (Spring/Summer/Autumn/Winter) */
  color_seasons: string[];
  /** Style archetypes this product suits */
  style_archetypes: string[];
  /** Calendar seasons this product is appropriate for */
  categories_calendar: string[];
  /** Affiliate-tracked URLs keyed by country code (US/GB/AU/CA/FR/DE) */
  affiliate_urls: ShopProductAffiliateUrls;
  /** How many user analyses have recommended this product */
  generated_count: number;
  created_at: Date;
  /**
   * How many of undertone/color_season/calendar_season matched (only in GET response)
   * @minimum 0
   * @maximum 3
   */
  match_score?: number;
}